Update ertIPManager.py (#726)
In ERTIPManager I added the option to run a pure ERT inversion. I think this is useful to be able to make sure that the resistivity model is good before running the IP inversions. Before my change, this was not possible, because the invert() function in ERTIPManager was overwriting the invert() function in ERTManager, and there is no way of constructing an ERTIPManager object directly from an ERTManager object. Creating an invertDC function seemed to me the simplest solution...